Move abstract repository into core namespace
