+++ src/Adapter/Shop/Repository/ShopGroupRepository.php
@@ -27,7 +27,7 @@ declare(strict_types=1);
 
 namespace PrestaShop\PrestaShop\Adapter\Shop\Repository;
 
-use PrestaShop\PrestaShop\Adapter\AbstractObjectModelRepository;
+use PrestaShop\PrestaShop\Core\Repository\AbstractObjectModelRepository;
 use PrestaShop\PrestaShop\Core\Domain\Shop\Exception\ShopGroupNotFoundException;
 use PrestaShop\PrestaShop\Core\Domain\Shop\ValueObject\ShopGroupId;
